move Server creation and start up in to consturctor method...
add message holding var for server, add program loop which takes a
message once connected, displays message on server side, and sends a
message back to client which is either a request to talk more followed
by a random number, or the first 5 chars of the recived message followed
by a random number.
when server recives texxt 'QUIT' the loop ends and the server shuts
down.